Segment Tree

+++ DSU.cpp
@@ -1,40 +1,52 @@
 #include <vector>
 using namespace std;
 
-class DSU {
+class DSU
+{
 public:
-    DSU(int N) {
+    DSU(long long N)
+    {
         prefixo.assign(N, 0);
         rank.assign(N, 0);
-        for (int i = 0; i < N; i++) {
+        for (long long i = 0; i < N; i++)
+        {
             prefixo[i] = i;
         }
     }
 
-    int FindSet(int i) {
+    long long FindSet(long long i)
+    {
         return prefixo[i] == i ? i : (prefixo[i] = FindSet(prefixo[i]));
     }
 
-    bool SameSet(int i, int j) {
+    bool SameSet(long long i, long long j)
+    {
         return FindSet(i) == FindSet(j);
     }
 
-    void UnionSet(int i, int j) {
-        if (!SameSet(i, j)) {
-            int a = FindSet(i);
-            int b = FindSet(j);
+    void UnionSet(long long i, long long j)
+    {
+        if (!SameSet(i, j))
+        {
+            long long a = FindSet(i);
+            long long b = FindSet(j);
 
-            if (rank[a] > rank[b]) {
+            if (rank[a] > rank[b])
+            {
                 prefixo[b] = a;
-            } else if (rank[b] > rank[a]) {
+            }
+            else if (rank[b] > rank[a])
+            {
                 prefixo[a] = b;
-            } else {
+            }
+            else
+            {
                 prefixo[b] = a;
                 rank[a]++;
             }
         }
     }
 
-    vector<int> prefixo;
-    vector<int> rank;
+    vector<long long> prefixo;
+    vector<long long> rank;
 };
